Fixing primitive value conversion warnings

+++ lib/DDHidAppleMikey.h
@@ -44,7 +44,7 @@
 
 - (NSArray *) pressElements;
 
-- (unsigned) numberOfKeys;
+- (unsigned long) numberOfKeys;
 
 - (void) addElementsToQueue: (DDHidQueue *) queue;
